apps/vanna/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
vanna[gpt]

